pyproject.toml
[tool.poetry]
name = "django-rest-scalable-apis-example"
version = "0.1.0"
description = ""
authors = ["Denis Orehovsky <denis.orehovsky@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
django-rest-batteries = "^1.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
